Fixed a small error in Strategy and added Strategy testing code to main

+++ SystemFiles/InorganicFertilizer.cpp
@@ -7,8 +7,8 @@ std::string InorganicFertilizer::getStrategyName() const{
 
 void InorganicFertilizer::applyFertilizer(Plant* p){
     std::cout << "applying the inorganic fertilizer strategy" << std::endl << "Current Nutrients level: " << p->getCurrentNutrients() << "ppm" <<std::endl;
-    std::cout << "Max Nutrients level: " << p->getMaxNutrients() << "ppm" << std::endl << "Nurtient level %: " << (p->getCurrentNutrients()/p->getMaxNutrients())*100 << std::endl;
+    std::cout << "Max Nutrients level: " << p->getMaxNutrients() << "ppm" << std::endl << "Nurtient level %: " << ((double)p->getCurrentNutrients()/p->getMaxNutrients())*100 << std::endl;
     std::cout << "Filling " << p->getName() << "with 40ppm of inorganic fertilizer" << std::endl;
     p->setCurrentNutrients(p->getCurrentNutrients() + 40 );
-    std::cout << "Current Nutrients level: " << p->getCurrentNutrients() << "ppm" <<std::endl << "Water level %: " << (p->getCurrentNutrients()/p->getMaxNutrients())*100 << std::endl;
+    std::cout << "Current Nutrients level: " << p->getCurrentNutrients() << "ppm" <<std::endl << "Water level %: " << ((double)p->getCurrentNutrients()/p->getMaxNutrients())*100 << std::endl;
 }
